Grass
Grass is "simulated" is a relatively unconventional way, using
heightmaps of high frequency perlin noise. This alone would create a
weak illusion of grass, as the field would look quite uniform. Thus, two
extra layers of perlin noise are added (which have lower frequency and
octaves): one for controlling the length of the blades of grass
(multiplying the height), and one for controller their width
(exponentiating the height; the heightmap stores values in [0, 1]).