Normal textures for water
This creates a more realistic and dynamic surface appearance for water
(by sliding multiple normal textures). In the future, however, we will
try to avoid using external textures and procedurally generate wave
textures, since something of this sort will anyways be needed for the
water to interact appropriately with the landscape.